unencrypted
I need to send the XPS-13 into Dell for service, as it seems to have a
buggy keyboard. To avoid confusion/complications during the repair
process, install without disk encryption.

This forced me to think a little harder about what exactly is happening
in the commands that create/format/mount volumes and filesystems.